Move all content from lib to Lib folder
